@@ -10,7 +10,7 @@ abstract class Kohana_HTTP_Exception extends Kohana_Exception {
 	 * @param   array   $variables  translation variables
 	 * @return  HTTP_Exception
 	 */
-	public static function factory($code, $message = NULL, array $variables = NULL, Throwable $previous = NULL)
+	public static function factory($code, $message = "", array $variables = NULL, Throwable $previous = NULL)
 	{
 		$class = 'HTTP_Exception_'.$code;
 
