@@ -60,7 +60,7 @@ class Kohana_Validation implements ArrayAccess {
 	 * @param   mixed    $value     value to set
 	 * @return  void
 	 */
-	public function offsetSet($offset, $value)
+	public function offsetSet($offset, $value): void
 	{
 		throw new Kohana_Exception('Validation objects are read-only.');
 	}
@@ -72,7 +72,7 @@ class Kohana_Validation implements ArrayAccess {
 	 * @param   string  $offset key to check
 	 * @return  bool    whether the key is set
 	 */
-	public function offsetExists($offset)
+	public function offsetExists($offset): bool
 	{
 		return isset($this->_data[$offset]);
 	}
@@ -85,7 +85,7 @@ class Kohana_Validation implements ArrayAccess {
 	 * @param   string  $offset key to unset
 	 * @return  void
 	 */
-	public function offsetUnset($offset)
+	public function offsetUnset($offset): void
 	{
 		throw new Kohana_Exception('Validation objects are read-only.');
 	}
@@ -97,7 +97,7 @@ class Kohana_Validation implements ArrayAccess {
 	 * @param   string  $offset key to return
 	 * @return  mixed   value from array
 	 */
-	public function offsetGet($offset)
+	public function offsetGet($offset): mixed
 	{
 		return $this->_data[$offset];
 	}
